create handle_redirects for execute_multple_commands, creates pipes.c
