+++ Makefile
@@ -10,7 +10,7 @@ MINIENV =	minienv.c minienv_utils.c minienv_str_utils.c free_minienv.c
 BUILTINS =	builtins_utils.c echo.c cd.c pwd.c export.c unset.c env.c exit.c
 EXECUTES =	execute_one_command.c execute_multiple_commands.c wait.c \
 			execute_external.c execute_builtin.c split_args.c get_path.c
-REDIRECTS =	redirects.c input_redirect.c file_descriptors.c
+REDIRECTS =	redirects.c input_redirect.c file_descriptors.c pipes.c
 UTILS =		error.c quote_checker.c signals.c str_utils.c \
 			ft_atoll.c free_array.c
 EXPANDS	= 	handle_expansions.c expand_variables.c expand_exit_status.c \
