+++ executes/wait.c
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/08 18:20:51 by sguilher          #+#    #+#             */
-/*   Updated: 2022/06/09 15:15:51 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/06/09 17:04:02 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
