+++ includes/executes.h
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 15:41:13 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/09 15:15:19 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/06/09 17:02:35 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,10 +35,13 @@ int		wait_for_children(int children_pid[1024]);
 // redirects
 void	redirect_fd(int fd_to_redirect, int fd_location);
 void	redirect_fds(int fd_in, int fd_out);
-void	prepare_io(int fd_stdout, int is_first_command, int has_next_command);
 int		redirect_input(char *command);
 char	*input_redirect_position(char *str);
 void	close_all_fds(void);
 void	close_extra_fds(void);
+int		has_input_redirect(char *command);
+
+// pipes 
+void	handle_pipe(int original_fd_out, char *current_command, char **commands);
 
 #endif
